pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voicecheck"
version = "0.1.0"
description = "Character Voice Consistency Checker"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "DaemonShip", email = "dev@daemonship.io"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-docx>=1.1.0",
    "anthropic>=0.7.0",
    "stripe>=7.8.0",
    "supabase>=2.3.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "sse-starlette>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mutmut>=2.4.0",
    "httpx>=0.25.0",
    "pre-commit>=3.5.0",
]

[tool.setuptools]
packages = ["backend", "backend.app", "backend.app.api", "backend.app.models", "backend.app.services"]
package-dir = {"" = "."}

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["backend"]
omit = ["tests/*"]

[tool.mutmut]
paths_to_mutate = ["backend"]
tests_dir = ["tests"]
